move transform call to report slice processor
