+++ yupana/processor/report_processor.py
@@ -44,8 +44,7 @@ from prometheus_client import Counter, Gauge
 
 from api.models import (Report, ReportSlice, Status)
 from api.serializers import ReportSerializer, ReportSliceSerializer
-from config.settings.base import (HOSTS_TRANSFORMATION_ENABLED,
-                                  INSIGHTS_KAFKA_ADDRESS,
+from config.settings.base import (INSIGHTS_KAFKA_ADDRESS,
                                   MAX_HOSTS_PER_REP,
                                   RETRIES_ALLOWED,
                                   RETRY_TIME)
@@ -186,8 +185,6 @@ class ReportProcessor(AbstractProcessor):  # pylint: disable=too-many-instance-a
             try:
                 self.report_json = json.loads(report_slice.report_json)
                 candidate_hosts = self._validate_report_details()
-                if HOSTS_TRANSFORMATION_ENABLED:
-                    candidate_hosts = self._transform_hosts(candidate_hosts)
 
                 if candidate_hosts:
                     self.status = SUCCESS_CONFIRM_STATUS
